Time: 0 ms (100%), Space: 23.9 MB (22.68%) - LeetHub

+++ 0637-average-of-levels-in-binary-tree/0637-average-of-levels-in-binary-tree.cpp
@@ -30,7 +30,7 @@ public:
 
                 sum+=node->val;
             }
-            ans.push_back((double)sum/(double)size);
+            ans.push_back(sum/size);
         }
 
         return ans;
